+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha35] - 2026-08-21
+
+### Changed
+
+- Justfile prod-up to use official images instead of local builds.
+
+### Fixed
+
+- Make `just prod-up` able to bring the stack up
+
 ## [Unreleased]
 
 ### Added